refactor(painting): Рефактор и добавление остальных фигур
