+++ src/main/java/com/example/paint/shapes/Circle.java
@@ -3,9 +3,10 @@ package com.example.paint.shapes;
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.control.ColorPicker;
 
-public class Circle extends Shape {
+public class Circle extends ShapeWithSize {
     @Override
     public void draw(GraphicsContext graphicsContext, ColorPicker colorPicker) {
-
+        super.draw(graphicsContext, colorPicker);
+        graphicsContext.strokeOval(startPoint.x, startPoint.y, width, height);
     }
 }
