+++ src/main/java/com/example/paint/shapes/Rectangle.java
@@ -3,9 +3,10 @@ package com.example.paint.shapes;
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.control.ColorPicker;
 
-public class Rectangle extends Shape {
+public class Rectangle extends ShapeWithSize {
     @Override
     public void draw(GraphicsContext graphicsContext, ColorPicker colorPicker) {
-
+        super.draw(graphicsContext, colorPicker);
+        graphicsContext.strokeRect(startPoint.x, startPoint.y, width, height);
     }
 }
