+++ src/main/java/com/example/paint/shapes/ShapeWithoutSize.java
@@ -1,0 +1,8 @@
+package com.example.paint.shapes;
+
+import javafx.scene.canvas.GraphicsContext;
+import javafx.scene.control.ColorPicker;
+
+public abstract class ShapeWithoutSize implements IShape {
+
+}
