--- src/main/java/com/example/paint/shapes/Triangle.java
@@ -1,11 +1,0 @@
-package com.example.paint.shapes;
-
-import javafx.scene.canvas.GraphicsContext;
-import javafx.scene.control.ColorPicker;
-
-public class Triangle extends Shape {
-    @Override
-    public void draw(GraphicsContext graphicsContext, ColorPicker colorPicker) {
-
-    }
-}
